rename: WhiteCard 컴포넌트 이름 index로 변경
